+++ example/src/components/form-builder-mui/elements/Checkboxes.tsx
@@ -1,6 +1,6 @@
 import React, { useState, ChangeEvent, FC } from "react";
 import { Typography, Checkbox, FormControlLabel, Box, FormControl, FormLabel, FormGroup, FormHelperText } from "@mui/material";
-import QuestionPromptText from "../components/QuestionPromptText";
+import QuestionPromptText from "../helperComponents/QuestionPromptText";
 
 export interface OptionType {
     value: string;
